refactor: renamed non-page component files to match component names
