+++ src/components/Posts/PostForm.tsx
@@ -15,8 +15,8 @@ import { useCreatePostMutation, useUpdatePostMutation } from "../../hooks/post";
 import { Post, PostsFormValues } from "../../types/post";
 import { generateRandom, redirectTo } from "../../utils/common";
 import { buildImageData } from "../../utils/image";
-import AttachedImages from "../Images/Attached";
-import ImageInput from "../Images/Input";
+import AttachedImages from "../Images/AttachedImages";
+import ImageInput from "../Images/ImageInput";
 import Flex from "../Shared/Flex";
 import Spinner from "../Shared/Spinner";
 import TextFieldWithAvatar from "../Shared/TextFieldWithAvatar";
