+++ src/components/Shared/TextFieldWithAvatar.tsx
@@ -1,6 +1,6 @@
 import { styled } from "@mui/material/styles";
 import { InputBase as MuiInputBase, InputBaseProps } from "formik-material-ui";
-import UserAvatar from "../Users/Avatar";
+import UserAvatar from "../Users/UserAvatar";
 import Flex from "./Flex";
 
 const InputBase = styled(MuiInputBase)<InputBaseProps>(({ theme }) => ({
