+++ src/pages/posts/[id]/edit.tsx
@@ -2,7 +2,7 @@ import { Button } from "@mui/material";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
 import { toastVar } from "../../../client/cache";
-import PostForm from "../../../components/Posts/Form";
+import PostForm from "../../../components/Posts/PostForm";
 import ProgressBar from "../../../components/Shared/ProgressBar";
 import { NavigationPaths } from "../../../constants/common";
 import { useTranslate } from "../../../hooks/common";
