+++ src/pages/users/[name].tsx
@@ -1,7 +1,7 @@
 import { Typography } from "@mui/material";
 import { NextPage } from "next";
 import { useRouter } from "next/router";
-import PostsList from "../../components/Posts/List";
+import PostList from "../../components/Posts/PostList";
 import ProgressBar from "../../components/Shared/ProgressBar";
 import ProfileCard from "../../components/Users/ProfileCard";
 import { useTranslate } from "../../hooks/common";
@@ -29,7 +29,7 @@ const UserProfile: NextPage = () => {
   return (
     <>
       <ProfileCard user={user} />
-      {user.posts && <PostsList posts={user.posts} />}
+      {user.posts && <PostList posts={user.posts} />}
     </>
   );
 };
